feat: User se muestran mas datos

+++ proyecto_final/src/app/modules/dashboard/pages/users/users.component.ts
@@ -1,44 +1,57 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 import { selectUsers, selectLoading, selectError } from './store/user.selectors';
 import { UsersService } from '../../../../core/services/users.service';
 import { User } from '../../../../interfaces/user';
+import { AuthService } from '../../../../core/services/auth.service';
 
 @Component({
   selector: 'app-users',
   standalone: false,
-
   templateUrl: './users.component.html',
   styleUrl: './users.component.scss'
 })
 
 export class UsersComponent implements OnInit, OnDestroy {
-  displayedColumns = ["name", "delete"];
+  displayedColumns: string[] = [];
   users$: Observable<User[]>;
   loading$: Observable<boolean>;
   error$: Observable<any>;
-  dataTableUsers: User[] = [];
+  isAdmin$: Observable<boolean>;
 
   constructor(
     private store: Store,
-    private usersService: UsersService
+    private usersService: UsersService,
+    private authService: AuthService
   ) {
     this.users$ = this.store.select(selectUsers);
     this.loading$ = this.store.select(selectLoading);
     this.error$ = this.store.select(selectError);
+    
+    this.isAdmin$ = this.authService.isAdmin$.pipe(
+      map(user => user?.role === 'ADMIN')
+    );
   }
 
   ngOnDestroy(): void {
-    this.usersService.resetUserState()
+    this.usersService.resetUserState();
   }
 
   ngOnInit(): void {
-    this.usersService.loadUsers();
+    this.usersService.loadUsers(); // Cargar usuarios siempre
+    
+    this.isAdmin$.subscribe(isAdmin => {
+      this.displayedColumns = isAdmin 
+        ? ['name', 'id', 'role', 'email', 'delete'] 
+        : ['name'];
+    });
   }
 
   deleteUserById(id: string) {
-    this.usersService.deleteUserById(id)
+    if (confirm('¿Estás seguro de eliminar este usuario?')) {
+      this.usersService.deleteUserById(id);
+    }
   }
-
 }
